Split ConnectorPrimitive and SerialConnectorPrimitive classes to separate files
